Change datatype of spatial cells mpi transfers to uint64_t

+++ spatial_cell.cpp
@@ -39,6 +39,6 @@ namespace spatial_cell {
    Real SpatialCell::cell_dvz = 0;
    Real SpatialCell::velocity_block_min_value = 0;    
    unsigned int SpatialCell::max_velocity_blocks = 0;
-   unsigned int SpatialCell::mpi_transfer_type = 0;
+   uint64_t SpatialCell::mpi_transfer_type = 0;
 }
